Adjust Helm prefix key
helm-command-prefix-key is already defined with defcustom in
helm-config.el. It must be set before helm-config.el to take
effect. With current setting, "C-x c" is used instead of "C-c h" and it
conflicts with the guide on homepage. "C-x c" makes it easy to press
"C-x C-c" to close Emacs when pressing fast enough. "C-c h" is also
originally used for Helm in Prelude.