User Specific save files
Prelude is generally stored in ~/.emacs.d of the user who wants to
use it. But for a system wide install of prelude it does not make
sense for every user on the system to install it and link to it.
It is better in that case to install it in a global position like
/opt/prelude and put the following line in site-start.el file

     (load "/opt/prelude/init.el")

Due to this change the local history and cache files of the
individual users get stored in their home directories instead
of trying to save it in prelude-dir